Message relay and refactors

+++ src/storage/MainStorage.sol
@@ -2,12 +2,12 @@
 pragma solidity ^0.8.22;
 
 abstract contract MainStorage {
-    uint64 s_txBatchId;
+    uint32 s_txBatchId;
 
-    mapping(uint64 => uint256) public s_txBatchId2StateRoot;
-    mapping(uint64 => uint256) public s_txBatchId2Timestamp;
+    mapping(uint32 => uint256) public s_txBatchId2StateRoot;
+    mapping(uint32 => uint256) public s_txBatchId2Timestamp;
 
-    mapping(uint64 => bool) s_accumulatedHashesRelayed; // txBatchId => wasRelayed
+    mapping(uint32 => bool) s_accumulatedHashesRelayed; // txBatchId => wasRelayed
 
     address s_L1MessageRelay; // The contract that passes messages from the L1 contract
     address s_escapeVerifier; // The contract that verifies the escape proofs
